Fix cmake warnings.
