+++ sbncode/BeamSpillInfoRetriever/ICARUSBNBRetriever/CMakeLists.txt
@@ -2,7 +2,7 @@ find_package(ifbeam)
 find_package(ifdh_art)
 
 cet_build_plugin(ICARUSBNBRetriever art::module
-    LIBRARIES
+    LIBRARIES PRIVATE
         sbn_POTTools
 	sbn_getFOM
         SQLite::SQLite3
