+++ sbncode/DetSim/CMakeLists.txt
@@ -31,8 +31,8 @@ set(	MODULE_LIBRARIES
 		FFTW3::FFTW3
 		)
 
-cet_build_plugin(AdjustSimForTrigger art::module LIBRARIES ${MODULE_LIBRARIES})
-cet_build_plugin(FilterSimEnergyDeposits art::module LIBRARIES ${MODULE_LIBRARIES})
+cet_build_plugin(AdjustSimForTrigger art::module LIBRARIES PRIVATE ${MODULE_LIBRARIES})
+cet_build_plugin(FilterSimEnergyDeposits art::module LIBRARIES PRIVATE ${MODULE_LIBRARIES})
 
 add_subdirectory(fcl)
