+++ sbncode/LArRecoProducer/CMakeLists.txt
@@ -28,18 +28,18 @@ set(  MODULE_LIBRARIES
         lardataobj::RecoBase
 )
 
-cet_build_plugin(CRTHitConverter art::module LIBRARIES ${MODULE_LIBRARIES})
-cet_build_plugin(MCSFitAllPID art::module LIBRARIES ${MODULE_LIBRARIES})
-cet_build_plugin(PMTFlashTriggerFilter art::module LIBRARIES ${MODULE_LIBRARIES})
-cet_build_plugin(PMTFlashTriggerMaker art::module LIBRARIES ${MODULE_LIBRARIES})
-cet_build_plugin(RangePAllPID art::module LIBRARIES ${MODULE_LIBRARIES})
-cet_build_plugin(ShowerCosmicDistance art::module LIBRARIES ${MODULE_LIBRARIES})
-cet_build_plugin(ShowerSelectionVars art::module LIBRARIES ${MODULE_LIBRARIES})
-cet_build_plugin(TrackScatterClosestApproach art::module LIBRARIES ${MODULE_LIBRARIES})
-cet_build_plugin(TrackStoppingChi2Fitter art::module LIBRARIES ${MODULE_LIBRARIES})
-cet_build_plugin(TransferPFParticleFlashMatch art::module LIBRARIES ${MODULE_LIBRARIES})
-cet_build_plugin(TransferPFParticleT0 art::module LIBRARIES ${MODULE_LIBRARIES})
-cet_build_plugin(TransferTrackT0 art::module LIBRARIES ${MODULE_LIBRARIES})
+cet_build_plugin(CRTHitConverter art::module LIBRARIES PRIVATE ${MODULE_LIBRARIES})
+cet_build_plugin(MCSFitAllPID art::module LIBRARIES PRIVATE ${MODULE_LIBRARIES})
+cet_build_plugin(PMTFlashTriggerFilter art::module LIBRARIES PRIVATE ${MODULE_LIBRARIES})
+cet_build_plugin(PMTFlashTriggerMaker art::module LIBRARIES PRIVATE ${MODULE_LIBRARIES})
+cet_build_plugin(RangePAllPID art::module LIBRARIES PRIVATE ${MODULE_LIBRARIES})
+cet_build_plugin(ShowerCosmicDistance art::module LIBRARIES PRIVATE ${MODULE_LIBRARIES})
+cet_build_plugin(ShowerSelectionVars art::module LIBRARIES PRIVATE ${MODULE_LIBRARIES})
+cet_build_plugin(TrackScatterClosestApproach art::module LIBRARIES PRIVATE ${MODULE_LIBRARIES})
+cet_build_plugin(TrackStoppingChi2Fitter art::module LIBRARIES PRIVATE ${MODULE_LIBRARIES})
+cet_build_plugin(TransferPFParticleFlashMatch art::module LIBRARIES PRIVATE ${MODULE_LIBRARIES})
+cet_build_plugin(TransferPFParticleT0 art::module LIBRARIES PRIVATE ${MODULE_LIBRARIES})
+cet_build_plugin(TransferTrackT0 art::module LIBRARIES PRIVATE ${MODULE_LIBRARIES})
 
 install_headers()
 install_fhicl()
