+++ sbncode/TPCReco/CalorimetryAnalysis/CMakeLists.txt
@@ -1,5 +1,5 @@
 cet_build_plugin( CalorimetryAnalysis art::module
-                        LIBRARIES
+                        LIBRARIES PRIVATE
                             art::Framework_Core
                             art::Framework_Services_Registry
                             art_root_io::TFileService_service
